requirements.txt
pandas
ta
python-binance
schedule
requests
python-telegram-bot
